requirements.txt
adal==1.2.7
aiohttp==3.8.4
aiosignal==1.3.1
aiounittest==1.4.2
appnope==0.1.3
arrow==1.2.3
asttokens==2.2.1
async-timeout==4.0.2
asyncio==3.4.3
attrs==23.1.0
azure-ai-language-conversations==1.0.0
azure-cognitiveservices-language-luis==0.2.0
azure-common==1.1.28
azure-core==1.27.0
azure-identity==1.13.0
azure-storage-blob==12.16.0
backcall==0.2.0
binaryornot==0.4.4
botbuilder-ai==4.14.4
botbuilder-core==4.14.4
botbuilder-schema==4.14.4
botframework-connector==4.14.4
botframework-streaming==4.14.4
certifi==2023.5.7
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.1.0
click==8.1.3
comm==0.1.3
cookiecutter==1.7.0
cryptography==3.4.8
debugpy==1.6.7
decorator==5.1.1
executing==1.2.0
frozenlist==1.3.3
future==0.18.3
idna==2.10
importlib-metadata==6.6.0
ipykernel==6.23.0
ipython==8.13.2
isodate==0.6.1
jedi==0.18.2
Jinja2==3.1.2
jinja2-time==0.2.0
jsonpickle==1.4.2
jupyter_client==8.2.0
jupyter_core==5.3.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
msal==1.22.0
msal-extensions==1.0.0
msrest==0.6.10
msrestazure==0.6.4
multidict==6.0.4
nest-asyncio==1.5.6
numpy==1.24.3
oauthlib==3.2.2
packaging==23.1
pandas==2.0.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.5.0
portalocker==2.7.0
poyo==0.5.0
prompt-toolkit==3.0.38
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.15.1
PyJWT==2.7.0
python-dateutil==2.8.2
pytz==2023.3
pyzmq==25.0.2
requests==2.25.1
requests-oauthlib==1.3.1
six==1.16.0
stack-data==0.6.2
tornado==6.3.1
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.5.0
tzdata==2023.3
ujson==5.7.0
urllib3==1.26.16
wcwidth==0.2.6
whichcraft==0.6.1
wrapt==1.15.0
yarl==1.9.2
zipp==3.15.0
